Separate SubCommand from Signing Engine (#264)
This was a refactor and resulted in no outward change in Ethsigner behaviour.
