+++ ethsigner/commandline/build.gradle
@@ -14,6 +14,7 @@ jar {
 dependencies {
 
   implementation project(':ethsigner:signing-api')
+  implementation project(':ethsigner:signer:common')
   implementation project(':ethsigner:core')
   implementation 'info.picocli:picocli'
   implementation 'com.google.guava:guava'
