+++ ethsigner/signer/common/src/main/java/tech/pegasys/ethsigner/TransactionSignerInitializationException.java
