+++ ethsigner/signer/multikey/build.gradle
@@ -27,6 +27,7 @@ dependencies {
 
   implementation project(':ethsigner:commandline')
   implementation project(':ethsigner:signing-api')
+  implementation project(':ethsigner:signer:common')
   implementation project(':ethsigner:signer:file-based')
   implementation project(':ethsigner:signer:azure')
   implementation project(':ethsigner:signer:hashicorp')
