+++ ethsigner/subcommands/build.gradle
@@ -1,18 +1,6 @@
-/*
- * Copyright 2018 ConsenSys AG.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- */
 
 jar {
-  archiveBaseName = 'ethsigner-signer-hashicorp'
+  archiveBaseName = 'ethsigner-subcommands'
   manifest {
     attributes(
       'Specification-Title': archiveBaseName,
@@ -23,33 +11,33 @@ jar {
   }
 }
 
-
 dependencies {
 
   implementation project(':ethsigner:commandline')
   implementation project(':ethsigner:signing-api')
+  implementation project(':ethsigner:core')
   implementation project(':ethsigner:signer:common')
+  implementation project(':ethsigner:signer:hashicorp')
   implementation project(':ethsigner:signer:file-based')
+  implementation project(':ethsigner:signer:multikey')
+  implementation project(':ethsigner:signer:azure')
+
   implementation 'tech.pegasys.signers.internal:hashicorp'
 
   implementation 'info.picocli:picocli'
   implementation 'com.google.guava:guava'
-  implementation 'org.apache.logging.log4j:log4j-api'
-  implementation 'org.web3j:core'
   implementation 'io.vertx:vertx-core'
-  implementation 'org.apache.tuweni:tuweni-net'
+  implementation 'org.web3j:core'
 
-  runtimeOnly 'org.bouncycastle:bcpkix-jdk15on'
-  runtimeOnly 'org.apache.logging.log4j:log4j-core'
+  implementation 'org.apache.logging.log4j:log4j-api'
+  implementation 'org.apache.logging.log4j:log4j-core'
   runtimeOnly 'org.apache.logging.log4j:log4j-slf4j-impl'
 
+  testImplementation project(path: ':ethsigner:commandline', configuration: 'testSupportArtifacts')
   testImplementation 'org.junit.jupiter:junit-jupiter-api'
   testImplementation 'org.junit.jupiter:junit-jupiter-params'
-  testImplementation 'io.vertx:vertx-codegen'
   testImplementation 'org.assertj:assertj-core'
-  testImplementation 'org.mockito:mockito-core'
-  testImplementation project(path: ':ethsigner:commandline', configuration: 'testSupportArtifacts')
-  testImplementation(group: 'tech.pegasys.signers.internal', name: 'hashicorp', classifier: 'test-fixtures')
+  testImplementation 'org.mockito:mockito-junit-jupiter'
 
   testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine'
 
@@ -57,4 +45,8 @@ dependencies {
   integrationTestImplementation 'org.assertj:assertj-core'
   integrationTestImplementation 'org.mock-server:mockserver-netty'
   integrationTestImplementation 'org.mockito:mockito-core'
+}
+
+artifacts {
+  testSupportArtifacts testSupportJar
 }
