+++ ethsigner/subcommands/src/main/java/tech/pegasys/ethsigner/subcommands/AzureSubCommand.java
@@ -10,16 +10,19 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package tech.pegasys.ethsigner.signer.azure;
+package tech.pegasys.ethsigner.subcommands;
 
 import static tech.pegasys.ethsigner.DefaultCommandValues.MANDATORY_PATH_FORMAT_HELP;
-import static tech.pegasys.ethsigner.util.PasswordFileUtil.readPasswordFromFile;
 
 import tech.pegasys.ethsigner.SignerSubCommand;
 import tech.pegasys.ethsigner.TransactionSignerInitializationException;
 import tech.pegasys.ethsigner.core.signing.SingleTransactionSignerProvider;
 import tech.pegasys.ethsigner.core.signing.TransactionSigner;
 import tech.pegasys.ethsigner.core.signing.TransactionSignerProvider;
+import tech.pegasys.ethsigner.signer.azure.AzureConfig;
+import tech.pegasys.ethsigner.signer.azure.AzureKeyVaultAuthenticator;
+import tech.pegasys.ethsigner.signer.azure.AzureKeyVaultTransactionSignerFactory;
+import tech.pegasys.ethsigner.util.PasswordFileUtil;
 
 import java.io.FileNotFoundException;
 import java.io.IOException;
@@ -77,7 +80,7 @@ public class AzureSubCommand extends SignerSubCommand {
   private TransactionSigner createSigner() throws TransactionSignerInitializationException {
     final String clientSecret;
     try {
-      clientSecret = readPasswordFromFile(clientSecretPath);
+      clientSecret = PasswordFileUtil.readPasswordFromFile(clientSecretPath);
     } catch (final FileNotFoundException fnfe) {
       throw new TransactionSignerInitializationException("File not found: " + clientSecretPath);
     } catch (final IOException e) {
