+++ ethsigner/subcommands/src/main/java/tech/pegasys/ethsigner/subcommands/FileBasedSubCommand.java
@@ -10,44 +10,59 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package tech.pegasys.ethsigner.signer.raw;
+package tech.pegasys.ethsigner.subcommands;
 
-import static tech.pegasys.ethsigner.DefaultCommandValues.MANDATORY_HOST_FORMAT_HELP;
+import static tech.pegasys.ethsigner.DefaultCommandValues.MANDATORY_FILE_FORMAT_HELP;
 
 import tech.pegasys.ethsigner.SignerSubCommand;
 import tech.pegasys.ethsigner.TransactionSignerInitializationException;
 import tech.pegasys.ethsigner.core.signing.SingleTransactionSignerProvider;
 import tech.pegasys.ethsigner.core.signing.TransactionSigner;
 import tech.pegasys.ethsigner.core.signing.TransactionSignerProvider;
-import tech.pegasys.ethsigner.signer.filebased.CredentialTransactionSigner;
+import tech.pegasys.ethsigner.signer.filebased.FileBasedSignerFactory;
+
+import java.nio.file.Path;
 
 import com.google.common.base.MoreObjects;
-import org.web3j.crypto.Credentials;
+import picocli.CommandLine;
 import picocli.CommandLine.Command;
 import picocli.CommandLine.Option;
+import picocli.CommandLine.Spec;
 
-/** Raw sub-command - DO NOT USE IN PRODUCTION, Private Key is in plain text */
+/** File-based authentication related sub-command */
 @Command(
-    name = RawSubCommand.COMMAND_NAME,
-    description =
-        "Sign transactions with an unecrypted key specified on cmdline. NOT SUITABLE FOR PRODUCTION.",
-    mixinStandardHelpOptions = true,
-    hidden = true)
-public class RawSubCommand extends SignerSubCommand {
+    name = FileBasedSubCommand.COMMAND_NAME,
+    description = "Sign transactions with a key stored in an encrypted V3 Keystore file.",
+    mixinStandardHelpOptions = true)
+public class FileBasedSubCommand extends SignerSubCommand {
 
-  static final String COMMAND_NAME = "raw-signer";
+  public static final String COMMAND_NAME = "file-based-signer";
+
+  public FileBasedSubCommand() {}
+
+  @SuppressWarnings("unused") // Picocli injects reference to command spec
+  @Spec
+  private CommandLine.Model.CommandSpec spec;
+
+  @Option(
+      names = {"-p", "--password-file"},
+      description = "The path to a file containing the password used to decrypt the keyfile.",
+      required = true,
+      paramLabel = MANDATORY_FILE_FORMAT_HELP,
+      arity = "1")
+  private Path passwordFilePath;
 
   @SuppressWarnings("FieldMayBeFinal") // Because PicoCLI requires Strings to not be final.
   @Option(
-      names = {"--key"},
-      description = "The private key used to generate signatures.",
-      paramLabel = MANDATORY_HOST_FORMAT_HELP,
+      names = {"-k", "--key-file"},
+      description = "The path to a file containing the key used to sign transactions.",
+      required = true,
+      paramLabel = MANDATORY_FILE_FORMAT_HELP,
       arity = "1")
-  private String privateKey;
+  private Path keyFilePath;
 
   private TransactionSigner createSigner() throws TransactionSignerInitializationException {
-    final Credentials credentials = Credentials.create(privateKey);
-    return new CredentialTransactionSigner(credentials);
+    return FileBasedSignerFactory.createSigner(keyFilePath, passwordFilePath);
   }
 
   @Override
@@ -63,6 +78,9 @@ public class RawSubCommand extends SignerSubCommand {
 
   @Override
   public String toString() {
-    return MoreObjects.toStringHelper(this).add("privateKey", privateKey).toString();
+    return MoreObjects.toStringHelper(this)
+        .add("passwordFilePath", passwordFilePath)
+        .add("keyFilePath", keyFilePath)
+        .toString();
   }
 }
