+++ settings.gradle
@@ -16,9 +16,11 @@ include 'acceptance-tests'
 include 'ethsigner:app'
 include 'ethsigner:core'
 include 'ethsigner:signing-api'
+include 'ethsigner:signer:common'
 include 'ethsigner:signer:hashicorp'
 include 'ethsigner:signer:file-based'
 include 'ethsigner:signer:azure'
 include 'ethsigner:signer:multikey'
-include 'ethsigner:signer:raw'
+include 'ethsigner:subcommands'
 include 'ethsigner:commandline'
+include 'signingservice'
